Removal of AFNetworking semi-final step: Deleted reference from Bridge Headers, AFRememberingSecurityPolicy.swift, OpenHABHTTPRequest.swift,  Moved Certificates class to own file Working without AFNetworking: Got rid of all warnings in Xcode! Added UIKit imports to ScaleAspectFitImageView, SegmentedUITableViewCell, SliderUITableViewCell, SwitchUITableViewCell
