+++ Podfile
@@ -5,7 +5,7 @@ use_frameworks!
 
 target 'openHAB' do
     #pod 'AFNetworking', '~> 3.0'
-    pod 'AFNetworking', '~> 2.6.0'
+    #pod 'AFNetworking', '~> 2.6.0'
     pod 'SDWebImage', '~> 5.0' , :modular_headers => true
     pod 'SDWebImageSVGCoder'
     pod 'GDataXML-HTML', '~> 1.3.0'
@@ -22,7 +22,7 @@ end
 target 'openHABTestsSwift' do
     inherit! :search_paths
     pod 'GDataXML-HTML', '~> 1.3.0'
-    pod 'AFNetworking', '~> 2.6.0'
+    #pod 'AFNetworking', '~> 2.6.0'
     #pod 'AFNetworking', '~> 3.0'
     pod 'Alamofire', '~> 4.0'
 end
