+++ openHAB-Bridging-Header.h
@@ -1,7 +1,7 @@
 //
 //  Use this file to import your target's public headers that you would like to expose to Swift.
 //
-#import <AFNetworking/AFNetworking.h>
+//#import <AFNetworking/AFNetworking.h>
 #import <GDataXML_HTML/GDataXMLNode.h>
 
 // -----------------------------------------------------------------------------
