+++ openHAB/OpenHABClientCertificatesViewController.swift
@@ -28,20 +28,20 @@ class OpenHABClientCertificatesViewController: UITableViewController {
 
     override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
         // Return the number of rows in the section.
-        return OpenHABHTTPRequestOperation.clientCertificateManager.clientIdentities.count
+        return NetworkConnection.clientCertificateManager.clientIdentities.count
     }
 
     static let tableViewCellIdentifier = "ClientCertificatesCell"
 
     override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
         let cell: UITableViewCell = tableView.dequeueReusableCell(withIdentifier: OpenHABClientCertificatesViewController.tableViewCellIdentifier, for: indexPath)
-        cell.textLabel?.text = OpenHABHTTPRequestOperation.clientCertificateManager.getIdentityName(index: indexPath.row)
+        cell.textLabel?.text = NetworkConnection.clientCertificateManager.getIdentityName(index: indexPath.row)
         return cell
     }
 
     override func tableView(_ tableView: UITableView, commit editingStyle: UITableViewCell.EditingStyle, forRowAt indexPath: IndexPath) {
         if editingStyle == UITableViewCell.EditingStyle.delete {
-            let status = OpenHABHTTPRequestOperation.clientCertificateManager.deleteFromKeychain(index: indexPath.row)
+            let status = NetworkConnection.clientCertificateManager.deleteFromKeychain(index: indexPath.row)
             if status == noErr {
                 tableView.deleteRows(at: [indexPath], with: .fade)
             }
