+++ openHAB/SwitchUITableViewCell.swift
@@ -8,6 +8,7 @@
 //  Converted to Swift 4 by Tim Müller-Seydlitz and Swiftify on 06/01/18
 //
 import os.log
+import UIKit
 
 fileprivate extension Selector {
     static let switchChange = #selector(SwitchUITableViewCell.switchChange)
