+++ openHABTests-Bridging-Header.h
@@ -2,6 +2,6 @@
 //  Use this file to import your target's public headers that you would like to expose to Swift.
 //
 
-#import <AFNetworking/AFNetworking.h>
+//#import <AFNetworking/AFNetworking.h>
 #import <GDataXML_HTML/GDataXMLNode.h>
