+++ openHABTestsSwift/OpenHABSecurityTests.swift
@@ -18,8 +18,6 @@ class OpenHABSecurityTests: XCTestCase {
         // Put teardown code here. This method is called after the invocation of each test method in the class.
     }
 
-    private var currentPageOperation: OpenHABHTTPRequestOperation?
-
     func testAFSecurity() {
       return
    }
